Refactor doctrine monitoring and database handling logic
Streamlined function structures, improved doctrine monitoring processes, and enhanced data handling for missing stats. Removed redundant code and logging, added new helper functions, and refined SQL integration for better performance and maintainability.
